mv logic from click context to api
